\name{palette.summary}
\alias{palette.summary}

\title{
Summary of all available palettes in the package.
}

\description{
Prints a summary of all available color palettes in the package categorized by artist and painting along with their numeric order.
}

\usage{
palette.summary()
}

\author{
Federico A Casale, \email{casale@hhu.de}
}

\seealso{
\link{see.all.paintings}, \link{see.palette}, \link{get.color},
}

\examples{
palette.summary()
}

