\name{see.all.paintings}
\alias{see.all.paintings}

\title{See all included paintings in the package.
}
\description{Displays a plot with all paintings included in the package.
}
\usage{
see.all.paintings()
}

\author{
Federico A Casale, \email{casale@hhu.de}
}

\seealso{
\link{palette.summary}, \link{see.palette}, \link{get.color}
}
\examples{
see.all.paintings()
}

